add userinput srv; add chair joints functions

+++ core/rammp_prototype_gui/rammp_prototype_gui/Gui_bridge.py
@@ -8,57 +8,107 @@ from std_msgs.msg import String
 from std_msgs.msg import Bool
 from .unreal_remote_websocket import UnrealRemoteWebsocket
 from .streaming.sender import StreamSender
-from gui_interfaces.msg import UserInputs
+from gui_interfaces.srv import UserInputs
 from sensor_msgs.msg import JointState
+from rclpy.callback_groups import ReentrantCallbackGroup
 
 
-class GuiMonitor(Node):
+class GuiBridge(Node):
     instance = None
 
-    def __init__(self, shm_size=1024 * 1024):
-        super().__init__("GuiMonitor")
-        GuiMonitor.instance = self
-        self.host = (
-            "192.168.1.21"  # TODO: make it configurable or load from config file
-        )
-        print("GuiMonitor node has been started.")
+    def __init__(
+        self,
+    ):
+        super().__init__("Gui_bridge")
+        print("Gui_bridge node has been started.")
+        GuiBridge.instance = self
 
-        self.ue = UnrealRemoteWebsocket(host=self.host, preset="RCPS")
+        self.declare_parameter("ue_host", "192.168.1.21")
+        self.host = self.get_parameter("ue_host").get_parameter_value().string_value
+
+        self.declare_parameter("use_shared_memory", False)
+        self.use_shared_memory = (
+            self.get_parameter("use_shared_memory").get_parameter_value().bool_value
+        )
+
+        self.declare_parameter("shm_size", 1024 * 1024)  # 1MB default size
+        self.shm_size = (
+            self.get_parameter("shm_size").get_parameter_value().integer_value
+        )
+
+        self.declare_parameter("ue_preset", "RCPS")
+        self.ue_preset = (
+            self.get_parameter("ue_preset").get_parameter_value().string_value
+        )
+        print(f"Using UE host: {self.host}, UE preset: {self.ue_preset}")
+
+        self.ue = UnrealRemoteWebsocket(host=self.host, preset=self.ue_preset)
         self.stream_sender = StreamSender(host=self.host, port=30030)
-        self.stream_timer = self.create_timer(0.1, self.send_stream_data)
+        self.stream_timer = self.create_timer(0.1, self.check_streamer_connection)
 
-        # Create shared memory and map it
-        self.shm = posix_ipc.SharedMemory(
-            "/ros_ue_shm",
-            posix_ipc.O_CREAT,
-            size=shm_size,  # default size is 1MB
-        )
-        # Map it
-        self.mapfile = mmap.mmap(self.shm.fd, shm_size)
-        self.shm.close_fd()
+        self._cb_group = ReentrantCallbackGroup()
+        self.arm_joints = None
+        self.base_joints = None
 
-        # make publisher for user connected or not, message should be boolen
-        self.connection_publisher = self.create_publisher(Bool, "user_connection", 1)
-        self.connection_publisher_timer = self.create_timer(
-            1.0, self.publish_connection_status
-        )
+        if self.use_shared_memory:
+            # Create shared memory and map it
+            self.shm = posix_ipc.SharedMemory(
+                "/ros_ue_shm",
+                posix_ipc.O_CREAT,
+                size=self.shm_size,  # default size is 1MB
+            )
+            # Map it
+            self.mapfile = mmap.mmap(self.shm.fd, self.shm_size)
+            self.shm.close_fd()
+
+        self.init_publisher()
+        self.init_subscriber()
+        self.init_service()
+
         self.test_ue_counter = 0
         self.test_ue_timer = self.create_timer(0.1, self.test_ue)
 
-        # make publisher for user input, message should be string
-        self.input_publisher = self.create_publisher(UserInputs, "user_input", 10)
+    def init_service(self):
+        # make service client for user input, request should be string
+        self.user_input_service_client = self.create_client(
+            UserInputs, "/GuiBridge/receive_input", callback_group=self._cb_group
+        )
 
+    def init_publisher(self):
+        # make publisher for user input, message should be string
+        self.connection_publisher = self.create_publisher(
+            Bool, "/GuiBridge/gui_connection", 10
+        )
+        self.connection_publisher_timer = self.create_timer(
+            1.0, self.publish_connection_status
+        )
+
+    def init_subscriber(self):
         # make subscriber for system state, message should be string
         self.system_state_subscriber = self.create_subscription(
-            String, "state", self.system_state_callback, 10
+            String,
+            "/system/state",
+            self.system_state_callback,
+            10,
+            callback_group=self._cb_group,
         )
         # make subscriber for arm joint state,
-        self.joint_state_subscriber = self.create_subscription(
-            JointState, "/arm/joint_states", self.joint_state_callback, 10
+        self.arm_joint_state_subscriber = self.create_subscription(
+            JointState,
+            "/arm/joint_states",
+            self.arm_joint_state_callback,
+            10,
+            callback_group=self._cb_group,
         )
-        self.arm_joints = [10.0] * 7  # Initialize with 7 joints
+        self.base_joint_state_subscriber = self.create_subscription(
+            JointState,
+            "/base/joint_states",
+            self.base_joint_state_callback,
+            10,
+            callback_group=self._cb_group,
+        )
 
-    def send_stream_data(self):
+    def check_streamer_connection(self):
         # check UE connection and connect/disconnect StreamSender accordingly
         if self.ue.is_connected() and not self.stream_sender.is_connected():
             try:
@@ -71,8 +121,11 @@ class GuiMonitor(Node):
         # if self.stream_sender.is_connected():
         #     print ("StreamSender is connected, sending stream data...")
 
-    def joint_state_callback(self, msg):
-        self.arm_joints = msg.position
+    def arm_joint_state_callback(self, msg):
+        self.arm_joints = msg
+
+    def base_joint_state_callback(self, msg):
+        self.base_joints = msg
 
     def test_ue(self):
         if self.ue.is_connected():
@@ -90,15 +143,47 @@ class GuiMonitor(Node):
     def set_ui_joints(self):
         arr = [0.0] * 53
         arm_joint_index_offset = 46  # index for link 0
-        for i in range(7):
-            arr[arm_joint_index_offset + i] = (
-                self.arm_joints[i] * 180.0 / 3.14159
-            )  # Convert radians to degrees
-        arr[arm_joint_index_offset + 0] = -arr[arm_joint_index_offset + 0]
-        arr[arm_joint_index_offset + 4] = -arr[arm_joint_index_offset + 4]
-        arr[arm_joint_index_offset + 5] = -arr[arm_joint_index_offset + 5]
-        arr[arm_joint_index_offset + 6] = -arr[arm_joint_index_offset + 6]
+        arm_gripper_joint_left_index = 38  # index for left gripper joint
+        arm_gripper_joint_right_index = 40  # index for right gripper joint
+        chair_fc_joint_index = 16  # index for chair foot control joint
+        chair_rc_joint_index = 25  # index for chair rotation control joint
+        chair_ml_joint_index = 1  # index for chair movement control joint
+        chair_mr_joint_index = 8  # index for chair movement control joint
+        chair_ml_carriage_index = 0  # index for chair movement carriage joint
+        chair_mr_carriage_index = 4  # index for chair movement carriage joint
+        chair_ml_wheel_index = 2  # index for chair movement wheel joint
+        chair_mr_wheel_index = 9  # index for chair movement wheel joint
+        if len(self.base_joints.position) >= 8:
+            for i in range(7):  # Only take the first 7 joints for the arm joint
+                arr[arm_joint_index_offset + i] = (
+                    self.arm_joints.position[i] * 180.0 / 3.14159
+                )  # Convert radians to degrees
+            arr[arm_joint_index_offset + 0] = -arr[arm_joint_index_offset + 0]
+            arr[arm_joint_index_offset + 4] = -arr[arm_joint_index_offset + 4]
+            arr[arm_joint_index_offset + 5] = -arr[arm_joint_index_offset + 5]
+            arr[arm_joint_index_offset + 6] = -arr[arm_joint_index_offset + 6]
+            # the 8th joint is the gripper, we can set it to 0 for now
+            # set gripper joint angle to 0 or 45 for close and open
+            arr[arm_gripper_joint_left_index] = (
+                0.0 if self.arm_joints.position[7] < 0.01 else -45.0
+            )
+            arr[arm_gripper_joint_right_index] = (
+                0.0 if self.arm_joints.position[7] < 0.01 else 45.0
+            )
+        if len(self.base_joints.position) >= 8:
+            arr[chair_fc_joint_index] = self.base_joints.position[0] * 180.0 / 3.14159
+            arr[chair_rc_joint_index] = self.base_joints.position[1] * 180.0 / 3.14159
+            arr[chair_mr_joint_index] = self.base_joints.position[2] * 180.0 / 3.14159
+            arr[chair_ml_joint_index] = self.base_joints.position[3] * 180.0 / 3.14159
+            arr[chair_ml_carriage_index] = self.base_joints.position[4] * 100.0  # to cm
+            arr[chair_mr_carriage_index] = self.base_joints.position[5] * 100.0  # to cm
+            arr[chair_ml_wheel_index] = self.base_joints.position[6] * 180.0 / 3.14159
+            arr[chair_mr_wheel_index] = self.base_joints.position[7] * 180.0 / 3.14159
+
         self.ue.call_function("setJoints", {"Values": arr})
+
+    def set_system_state(self):
+        self.ue.call_function("setSystemState", {"Values": self._system_state})
 
     def publish_connection_status(self):
         msg = Bool()
@@ -115,24 +200,27 @@ class GuiMonitor(Node):
         start = time.time()
         while self.loop.is_running() and (time.time() - start) < timeout:
             time.sleep(0.1)
-        self.mapfile.close()
-        self.shm.unlink()
+        if self.use_shared_memory:
+            self.mapfile.close()
+            self.shm.unlink()
         super().destroy_node()
 
     def system_state_callback(self, msg):
         state = msg.data
         print(f"Received system state: {state}")
+        self._system_state = msg.data
 
     def write_data_to_shm(self, index, data):
-        # Write data to shared memory
-        self.mapfile.seek(index)
-        self.mapfile.write(data.encode("utf-8"))
-        self.mapfile.flush()
+        if self.use_shared_memory:
+            # Write data to shared memory
+            self.mapfile.seek(index)
+            self.mapfile.write(data.encode("utf-8"))
+            self.mapfile.flush()
 
 
 def main():
     rclpy.init()
-    node = GuiMonitor()
+    node = GuiBridge()
     rclpy.spin(node)
     node.destroy_node()
     rclpy.shutdown()
